Add non-static XMLStreamParserComponent.isParseTarget(String name) method so that subclasses can override parse target. New DefaultGroup (subclass of ReachGroup) represents model-wide adjustments.  PredictContext.xml chg'ed to include, new tests, added as field of AdjustmentGroups. Chg ModelID to a Long, which affects several classes Add PredictionContext full constructor so instances can be build for test. Rm PredictionContext setModelID method - use full constructor (affects a few classes)

+++ src/gov/usgswim/sparrow/cachefactory/ReachByPointFactory.java
@@ -34,7 +34,7 @@ public class ReachByPointFactory extends AbstractCacheFactory {
 		
 		IDByPointRequest2 req = (IDByPointRequest2) request;
 		
-		Integer modelId = req.getModelID();
+		Long modelId = req.getModelID();
 		Double lng = req.getPoint().x;
 		Double lat = req.getPoint().y;
